requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-monitoring==0.31.1
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyMySQL==0.9.3
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
